Move field management to different module
