+++ server/api/teams/index.ts
@@ -1,7 +1,7 @@
 import db from './../../../init/database';
-import { Team, CustomFieldEntry, CustomFieldStore } from '../../../types/interfaces';
+import { Team } from '../../../types/interfaces';
 
-const { teams, custom } = db;
+const { teams } = db;
 
 export async function getTeamById(id: string, logo = false): Promise<Team | null> {
 	return new Promise(res => {
@@ -24,50 +24,3 @@ export const getTeamsList = (query: any) =>
 			return res(teams);
 		});
 	});
-
-export const initiateCustomFields = () =>
-	new Promise<CustomFieldStore>(res => {
-		custom.findOne({}, (err, store) => {
-			if (store) {
-				return res(store);
-			}
-			const customFields = { players: [], teams: [] };
-			custom.insert(customFields, (err, entry) => {
-				return res(entry);
-			});
-		});
-	});
-
-export const getTeamFields = async () => {
-	const store = await initiateCustomFields();
-	if (!store) return [];
-	return store.teams;
-};
-
-export const updateTeamFields = async (teamFields: CustomFieldEntry[]) => {
-	const store = await initiateCustomFields();
-
-	const deletedFields = store.teams.filter(field => !teamFields.find(newField => newField.name === field.name));
-	const createdFields = teamFields.filter(newField => !store.teams.find(field => field.name === newField.name));
-
-	return new Promise<CustomFieldStore>(res => {
-		custom.update({}, { $set: { teams: teamFields } }, { multi: true }, async () => {
-			if (!deletedFields.length && !createdFields.length) {
-				return res(await initiateCustomFields());
-			}
-			const updateQuery = {
-				$unset: {},
-				$set: {}
-			};
-			for (const deletedField of deletedFields) {
-				updateQuery.$unset[`extra.${deletedField.name}`] = true;
-			}
-			for (const createdField of createdFields) {
-				updateQuery.$set[`extra.${createdField.name}`] = '';
-			}
-			teams.update({}, updateQuery, { multi: true }, async () => {
-				res(await initiateCustomFields());
-			});
-		});
-	});
-};
